fixed packages structure
