--- src/main/java/com/example/matchmaker/control/mapping/UserMapper.java
@@ -1,9 +1,0 @@
-package com.example.matchmaker.control.mapping;
-
-import com.example.matchmaker.boundary.dto.UserDto;
-import com.example.matchmaker.entity.User;
-
-public interface UserMapper {
-
-    User from(UserDto dto);
-}
