--- src/main/java/com/example/matchmaker/control/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.example.matchmaker.control.repository;
-
-import com.example.matchmaker.entity.User;
-
-import java.util.List;
-
-public interface UserRepository {
-
-    void save(User user);
-
-    User findBy(String name);
-
-    List<User> findAll();
-
-    User deleteBy(String name);
-}
