--- src/main/java/com/example/matchmaker/control/service/MatchmakerService.java
@@ -1,10 +1,0 @@
-package com.example.matchmaker.control.service;
-
-import com.example.matchmaker.boundary.dto.UserDto;
-
-public interface MatchmakerService {
-
-    void accept(UserDto userDto);
-
-    void matchUsers();
-}
