--- src/main/java/com/example/matchmaker/control/service/UserGroupFactory.java
@@ -1,10 +1,0 @@
-package com.example.matchmaker.control.service;
-
-import com.example.matchmaker.entity.User;
-
-import java.util.List;
-
-public interface UserGroupFactory {
-
-    void createGroup(List<User> users);
-}
