--- src/main/java/com/example/matchmaker/control/service/UserGroupHandler.java
@@ -1,8 +1,0 @@
-package com.example.matchmaker.control.service;
-
-import com.example.matchmaker.entity.UserGroup;
-
-public interface UserGroupHandler {
-
-    void handleUserGroup(UserGroup userGroup);
-}
