+++ src/main/java/com/example/matchmaker/dto/UserDto.java
@@ -1,0 +1,23 @@
+package com.example.matchmaker.dto;
+
+import lombok.Builder;
+import lombok.Value;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Positive;
+import javax.validation.constraints.PositiveOrZero;
+
+@Value
+@Builder
+public class UserDto {
+
+    @NotBlank
+    String name;
+    @NotNull
+    @PositiveOrZero
+    Double skill;
+    @NotNull
+    @Positive
+    Double latency;
+}
