+++ src/main/java/com/example/matchmaker/mapping/UserMapper.java
@@ -1,0 +1,9 @@
+package com.example.matchmaker.mapping;
+
+import com.example.matchmaker.dto.UserDto;
+import com.example.matchmaker.entity.User;
+
+public interface UserMapper {
+
+    User from(UserDto dto);
+}
