+++ src/main/java/com/example/matchmaker/service/MatchmakerService.java
@@ -1,0 +1,10 @@
+package com.example.matchmaker.service;
+
+import com.example.matchmaker.dto.UserDto;
+
+public interface MatchmakerService {
+
+    void accept(UserDto userDto);
+
+    void matchUsers();
+}
