+++ src/main/java/com/example/matchmaker/service/UserGroupFactory.java
@@ -1,0 +1,10 @@
+package com.example.matchmaker.service;
+
+import com.example.matchmaker.entity.User;
+
+import java.util.List;
+
+public interface UserGroupFactory {
+
+    void createGroup(List<User> users);
+}
