+++ src/main/java/com/example/matchmaker/service/UserGroupHandler.java
@@ -1,0 +1,8 @@
+package com.example.matchmaker.service;
+
+import com.example.matchmaker.entity.UserGroup;
+
+public interface UserGroupHandler {
+
+    void handleUserGroup(UserGroup userGroup);
+}
