+++ src/test/java/com/example/matchmaker/util/UserDtoFactory.java
@@ -1,6 +1,6 @@
 package com.example.matchmaker.util;
 
-import com.example.matchmaker.boundary.dto.UserDto;
+import com.example.matchmaker.dto.UserDto;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AccessLevel;
